change DetailMeeting Activity into Fragment
