+++ app/src/main/res/layout-land/activity_meeting_details.xml
@@ -70,13 +70,6 @@
                     android:text="" />
 
                 <TextView
-                    android:id="@+id/meetingDetails_shadowPlaceName"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:visibility="gone"
-                    android:text="" />
-
-                <TextView
                     android:id="@+id/meetingDetails_subject"
                     android:layout_width="match_parent"
                     android:layout_height="wrap_content"
