+++ app/src/main/res/values/strings.xml
@@ -61,4 +61,6 @@
     <string name="date">date</string>
     <string name="meetings_filters">Meetings Filters</string>
     <string name="meeting_id">Meeting ID :</string>
+    <!-- TODO: Remove or change this placeholder text -->
+    <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
